try moving customization to external file
